Moved `Section` to core
