+++ app/src/main/java/com/imashnake/animite/features/home/HomeViewModel.kt
@@ -8,10 +8,10 @@ import com.imashnake.animite.api.anilist.type.MediaSort
 import com.imashnake.animite.api.anilist.type.MediaType
 import com.imashnake.animite.core.data.Resource
 import com.imashnake.animite.core.data.Resource.Companion.asResource
+import com.imashnake.animite.core.ui.models.Section
 import com.imashnake.animite.dev.ext.nextSeason
 import com.imashnake.animite.dev.ext.season
 import com.imashnake.animite.dev.internal.Constants
-import com.imashnake.animite.dev.internal.Section
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.SharingStarted
