+++ core/src/main/kotlin/com/imashnake/animite/core/ui/models/Section.kt
@@ -1,6 +1,7 @@
-package com.imashnake.animite.dev.internal
+package com.imashnake.animite.core.ui.models
 
 data class Section<T>(
     val header: String?,
     val body: T
 )
+
